Upgrade to PySide6

+++ src/opencmiss/zincwidgets/basesceneviewerwidget.py
@@ -13,7 +13,7 @@ file, You can obtain one at http://mozilla.org/MPL/2.0/.
 # This python module is intended to facilitate users creating their own applications that use OpenCMISS-Zinc
 # See the examples at https://svn.physiomeproject.org/svn/cmiss/zinc/bindings/trunk/python/ for further
 # information.
-from PySide2 import QtCore, QtWidgets
+from PySide6 import QtCore, QtWidgets
 
 from opencmiss.zinc.sceneviewer import Sceneviewer, Sceneviewerevent
 from opencmiss.zinc.sceneviewerinput import Sceneviewerinput
@@ -317,7 +317,7 @@ class BaseSceneviewerWidget(QtWidgets.QOpenGLWidget, InteractionManager):
         """
         Respond to widget resize events.
         """
-        self._sceneviewer.setViewportSize(width * self._pixel_scale, height * self._pixel_scale)
+        self._sceneviewer.setViewportSize(int(width * self._pixel_scale), int(height * self._pixel_scale))
 
     def keyPressEvent(self, event):
         self.key_press_event(event)
